As the Sink class is more globally used than originally planned (now also for the cachedigestion besides the piperunner), the class with subclasses is moved to its own package, thrown exceptions were made more generic and the Sink now throws an IOException (as it is used within Runnables and without Runnables). Runnables now need to catch these IOExceptions and caast it to a UncheckedIOException (custom class as UncheckedIOException is only introduced since Java8).
