+++ hadoop-pipeline-application/src/main/java/org/molgenis/hadoop/pipeline/application/inputstreamdigestion/Sink.java
@@ -1,7 +1,9 @@
-package org.molgenis.hadoop.pipeline.application.processes;
+package org.molgenis.hadoop.pipeline.application.inputstreamdigestion;
 
 import java.io.IOException;
 import java.io.InputStream;
+
+import org.molgenis.hadoop.pipeline.application.processes.PipeRunner;
 
 /**
  * Abstract class describing required methods for a {@link PipeRunner}{@code 's} last {@link Process}{@code '}
@@ -16,8 +18,9 @@ public abstract class Sink<T>
 	 * 
 	 * @param inputStream
 	 *            {@link InputStream}
+	 * @throws IOException
 	 */
-	public abstract void handleInputStream(InputStream inputStream);
+	public abstract void handleInputStream(InputStream inputStream) throws IOException;
 
 	/**
 	 * Digests a single {@code <T>item} from the {@link InputStream} digested by {@link #handleInputStream(InputStream)}
@@ -26,6 +29,7 @@ public abstract class Sink<T>
 	 * 
 	 * @param item
 	 *            {@code <T>}
+	 * @throws IOException
 	 */
 	protected abstract void digestStreamItem(T item) throws IOException;
 }
