+++ hadoop-pipeline-application/src/main/java/org/molgenis/hadoop/pipeline/application/inputstreamdigestion/StringSink.java
@@ -1,4 +1,4 @@
-package org.molgenis.hadoop.pipeline.application.processes;
+package org.molgenis.hadoop.pipeline.application.inputstreamdigestion;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,7 +6,6 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 
 import org.apache.commons.io.IOUtils;
-import org.molgenis.hadoop.pipeline.application.exceptions.SinkIOException;
 
 /**
  * Sink for digesting input streams line-by-line.
@@ -18,7 +17,7 @@ public abstract class StringSink extends Sink<String>
 	 * {@link #digestStreamItem(String)} is called.
 	 */
 	@Override
-	public void handleInputStream(InputStream inputStream)
+	public void handleInputStream(InputStream inputStream) throws IOException
 	{
 		BufferedReader br = null;
 		try
@@ -29,10 +28,6 @@ public abstract class StringSink extends Sink<String>
 			{
 				digestStreamItem(line);
 			}
-		}
-		catch (IOException e)
-		{
-			throw new SinkIOException(e);
 		}
 		finally
 		{
