+++ hadoop-pipeline-application/src/main/java/org/molgenis/hadoop/pipeline/application/processes/PipeInHandler.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 import java.io.OutputStream;
 
 import org.apache.commons.io.IOUtils;
-import org.molgenis.hadoop.pipeline.application.exceptions.ProcessPipeException;
+import org.molgenis.hadoop.pipeline.application.exceptions.UncheckedIOException;
 
 /**
  * Writes the data stored in a {@code byte[]} to the {@link OutputStream}.
@@ -47,7 +47,7 @@ public class PipeInHandler implements Runnable
 		}
 		catch (IOException e)
 		{
-			throw new ProcessPipeException(e);
+			throw new UncheckedIOException(e);
 		}
 		finally
 		{
