+++ hadoop-pipeline-application/src/main/java/org/molgenis/hadoop/pipeline/application/processes/PipeRunner.java
@@ -7,7 +7,8 @@ import java.io.InputStream;
 import java.io.OutputStream;
 
 import org.apache.commons.io.IOUtils;
-import org.molgenis.hadoop.pipeline.application.exceptions.ProcessPipeException;
+import org.molgenis.hadoop.pipeline.application.exceptions.UncheckedIOException;
+import org.molgenis.hadoop.pipeline.application.inputstreamdigestion.Sink;
 
 /**
  * Class for running a pipe of one or more {@link Process}{@code es}.
@@ -83,7 +84,7 @@ public class PipeRunner implements Runnable
 		}
 		catch (InterruptedException e)
 		{
-			throw new ProcessPipeException(e);
+			throw new RuntimeException(e);
 		}
 	}
 
@@ -99,7 +100,7 @@ public class PipeRunner implements Runnable
 		}
 		catch (IOException e)
 		{
-			throw new ProcessPipeException(e);
+			throw new UncheckedIOException(e);
 		}
 		finally
 		{
