+++ hadoop-pipeline-application/src/test/java/org/molgenis/hadoop/pipeline/application/processes/PipeRunnerTester.java
@@ -7,6 +7,8 @@ import org.apache.log4j.Logger;
 import org.apache.log4j.SimpleLayout;
 import org.apache.log4j.WriterAppender;
 import org.molgenis.hadoop.pipeline.application.Tester;
+import org.molgenis.hadoop.pipeline.application.inputstreamdigestion.SamRecordSink;
+import org.molgenis.hadoop.pipeline.application.inputstreamdigestion.StringSink;
 import org.testng.Assert;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
